requirements.txt
python-binance
pandas
numpy
TA-Lib
python-dotenv
vectorbt
matplotlib
plotly
requests
groq
